Remove TransportUris
