+++ app/core/src/main/java/com/fsck/k9/preferences/SettingsExporter.java
@@ -28,7 +28,6 @@ import com.fsck.k9.Preferences;
 import com.fsck.k9.backend.BackendManager;
 import com.fsck.k9.helper.FileHelper;
 import com.fsck.k9.mail.ServerSettings;
-import com.fsck.k9.mail.TransportUris;
 import com.fsck.k9.preferences.Settings.InvalidSettingValueException;
 import com.fsck.k9.preferences.Settings.SettingsDescription;
 import org.xmlpull.v1.XmlSerializer;
@@ -268,7 +267,7 @@ public class SettingsExporter {
 
 
         // Write outgoing server settings
-        ServerSettings outgoing = TransportUris.decodeTransportUri(account.getTransportUri());
+        ServerSettings outgoing = backendManager.decodeTransportUri(account.getTransportUri());
         serializer.startTag(null, OUTGOING_SERVER_ELEMENT);
         serializer.attribute(null, TYPE_ATTRIBUTE, outgoing.type.name());
